requirements.txt
pandas>=0.20
numpy>=1.11.1
requests>=2.11.1
six>=1.10.0
python-dateutil>=2.5.3
beautifulsoup4

# for Python 2.7 and 2.6 (environment marker support as of pip 6.0)
futures ; python_version < '3.3'
mock ; python_version < '3.3'

unittest2 ; python_version < '2.7'
ordereddict ; python_version < '2.7'


